Delete empty <context>/embeddings folders
This is a leftover from when we distinguished between "global" documents and context-specific documents. This is not really relevant for the user anymore. While it's still supported to load those subfolders from older knowledge packs that might still have that structure, I'm deleting these folders in this sample to avoid confusion