Adding threat modelling for "cards" interaction
With the newest Haiven version, it is possible to put in prompts that are used in the "cards" interaction mode, i.e. the prompt needs to ask for JSON that returns objects with at least "title" and "summary" properties, and then it will show those as a series of cards, instead of going into a text chat mode.

This feature is not really documented yet and a bit advanced, but here is an example.

It will also only work with models that are reliable in actually returning JSON, e.g., it usually doesn't work for us when we use a local model in Ollama, because most of them are too small to be able to do that.